Done with the tutorial. Will ad some additional functionalities in some time

+++ pages/api/products/[id].js
@@ -5,9 +5,10 @@ export default async function handler(req, res) {
   const {
     method,
     query: { id },
+    cookies,
   } = req;
+  const token = cookies.token;
   await dbConnect();
-
   if (method === "GET") {
     try {
       const product = await Product.findById(id);
@@ -18,6 +19,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "PUT") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("You are not authenticated");
+    }
     try {
       const product = await Product.create(req.body);
       res.status(200).json(product);
@@ -27,6 +31,9 @@ export default async function handler(req, res) {
   }
 
   if (method === "DELETE") {
+    if (!token || token !== process.env.TOKEN) {
+      return res.status(401).json("You are not authenticated");
+    }
     try {
       await Product.findByIdAndDelete(id);
       res.status(200).json("Product has been deleted!");
